refactor: refactored mistral specific functions to a separate folder
